Add multi-user install option.
When we are running as user 0, the install script will assume that we
want a "multi-user install". This means that multiple users can access
/nix through the "nix-daemon" service.

This will create users for a multi-user install using either OS X's
dseditgroup and dscl or the more common Unix groupadd and useradd.
This still needs to be tested to see whether the script breaks when
one or more of the group/users already exists.

This will also setup the permissions in /nix so that they are
"root:nixbld" 0755. This prevents ordinary users from modifying
/nix/store/, etc. "build-users-group" is set to nixbld automatically.

*More testing of multi-user mode is still needed.*